Some reorg in platform enums, using enum for GameConfig::Window::Mode
