+++ src/Application/GameConfig.cpp
@@ -15,6 +15,19 @@ MM_DEFINE_ENUM_SERIALIZATION_FUNCTIONS(RendererType, CASE_INSENSITIVE, {
     {RendererType::DirectDraw, "DirectDraw"}
 })
 
+MM_DEFINE_ENUM_SERIALIZATION_FUNCTIONS(PlatformWindowMode, CASE_INSENSITIVE, {
+    {WINDOW_MODE_WINDOWED, "windowed"},
+    {WINDOW_MODE_BORDERLESS, "borderless"},
+    {WINDOW_MODE_FULLSCREEN, "fullscreen"},
+    {WINDOW_MODE_FAKE_FULLSCREEN, "fake_fullscreen"},
+
+    // Compatibility with old serialization:
+    {WINDOW_MODE_WINDOWED, "0"},
+    {WINDOW_MODE_BORDERLESS, "1"},
+    {WINDOW_MODE_FULLSCREEN, "2"},
+    {WINDOW_MODE_FAKE_FULLSCREEN, "3"}
+})
+
 using EngineIoc = Engine_::IocContainer;
 using Application::GameConfig;
