+++ src/GUI/GUIWindow.h
@@ -12,7 +12,7 @@
 
 #include "GUI/GUIDialogues.h"
 
-#include "Platform/PlatformKey.h"
+#include "Platform/PlatformEnums.h"
 #include "Io/Mouse.h"
 #include "Io/InputAction.h"
 #include "Io/KeyboardInputHandler.h"
